packages/pdfsys-parser-pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfsys-parser-pipeline"
version = "0.0.1"
description = "OCR pipeline backend: consumes cached LayoutDocument + region-level OCR engine (RapidOCR / PaddleOCR-classic)."
requires-python = ">=3.11"
dependencies = [
    "pdfsys-core",
    "pymupdf>=1.25",
    "rapidocr-onnxruntime>=1.3",
    "numpy>=1.24",
    "Pillow>=10.0",
]

[tool.uv.sources]
pdfsys-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/pdfsys_parser_pipeline"]
